lvrend: Split border code into separate file
File split with no logic changes.
Move DrawBorder(), DrawBackgroundImage(), and DrawBodyBackground() out of
lvrend.cpp into their own drawutil.cpp/drawutil.h, as the drawing details
of border/background painting have no bearing on how DrawDocument() walks
the tree.
